Final html structure for a recipes webpage
Webpage about three diferent recipes, the html structure contains an index file wth all the links to the different html pages that contain the recipes. The pages contain pretty much the same structure, a main tittle with the name of the food recipe, then an image with that food along with a short description of the recipe in two paragraphs, and finally an ingredients and steps lists.